+++ src/Module/AddSvgSupport.php
@@ -21,14 +21,14 @@ class AddSvgSupport extends Instance
 {
     public function run()
     {
-        $this->setup();
-        $this->addSvgSupport();
+        $this->setup()->addSvgSupport();
     }
 
     protected function setup()
     {
         $this->setDefaultConfig(['admin', 'editor', 'author']);
         $this->config = $this->aliasUserRoles($this->config);
+        return $this;
     }
 
     public function addSvgSupport()
@@ -39,7 +39,28 @@ class AddSvgSupport extends Instance
                     $mimes['svg'] = 'image/svg+xml';
                     return $mimes;
                 });
+                $this->addSvgBugFix();
             }
         }
+    }
+    
+    protected function addSvgBugFix() {
+        // WordPress v4.7.1+ 
+        // http://codepen.io/chriscoyier/post/wordpress-4-7-1-svg-upload
+        add_filter( 'wp_check_filetype_and_ext', function($data, $file, $filename, $mimes) {
+            global $wp_version;
+            if ($wp_version === '4.7' || ((float) $wp_version < 4.7 )) {
+                return $data;
+            }
+
+            $filetype = wp_check_filetype( $filename, $mimes );
+
+            return [
+                'ext'             => $filetype['ext'],
+                'type'            => $filetype['type'],
+                'proper_filename' => $data['proper_filename']
+            ];
+
+        }, 10, 4 );
     }
 }